+++ NotificationService.py
@@ -1,8 +1,18 @@
-				mail.sendemail(from_addr    = 'fenaldy12@gmail.com', 
-        				to_addr_list = ['fenaldy13@gmail.com'],
-        				cc_addr_list = ['example@gmail.com'], 
-        				subject      = 'Pesan Error !', 
-        				message      =  msg, 
-        				login        = 'fenaldy12@gmail.com', 
-        				password     = 'mefor345')
-	
+
+import smtplib
+
+def sendemail(from_addr, to_addr_list, cc_addr_list,
+              subject, message,
+              login, password,
+              smtpserver='smtp.gmail.com', smtpport=587):  # split smtpserver and -port
+    header  = 'From: %s\n' % from_addr
+    header += 'To: %s\n' % ','.join(to_addr_list)
+    header += 'Cc: %s\n' % ','.join(cc_addr_list)
+    header += 'Subject: %s\n\n' % subject
+    message = header + message
+
+    server = smtplib.SMTP(smtpserver, smtpport)  # use both smtpserver  and -port 
+    server.starttls()
+    server.login(login,password)
+    problems = server.sendmail(from_addr, to_addr_list, message)
+    server.quit()
