--- err_template.txt
@@ -1,33 +1,0 @@
-Zend_Session::start() - /var/www/ABIYASA_1_1/library/Zend/Session.php(Line:482): Error #8 session_start(): ps_files_cleanup_dir: opendir(/var/lib/php5/sessions) failed: Permission denied (13)
-file: /var/www/ABIYASA_1_1/library/Zend/Session.php
-Array
-(
-    [0] => Array
-        (
-            [file] => /var/www/ABIYASA_1_1/library/Zend/Session/Namespace.php
-            [line] => 143
-            [function] => start
-            [class] => Zend_Session
-            [type] => ::
-            [args] => Array
-                (
-                    [0] => 1
-                )
-
-        )
-
-    [1] => Array
-        (
-            [file] => /var/www/ABIYASA_1_1/index.php
-            [line] => 67
-            [function] => __construct
-            [class] => Zend_Session_Namespace
-            [type] => ->
-            [args] => Array
-                (
-                    [0] => api.ns
-                )
-
-        )
-
-)
